+++ examples/turnstile/basic/turnstile.go
@@ -8,8 +8,8 @@ type Turnstile struct {
 	state string
 }
 
-// State returns the current state of the turnstile.
-func (t *Turnstile) State() string {
+// GetState returns the current state of the turnstile.
+func (t *Turnstile) GetState() string {
 	return t.state
 }