+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
 
+## Unreleased
+
+### Changed
+
+- Renamed `Subject.State` to `Subject.GetState` to avoid name collisions with `State` properties.
+
+
 ## 0.1.0 - 2017-08-28
 
 - Initial release